+++ common/native/apsi/result_package.h
@@ -13,15 +13,15 @@ namespace apsi
      */
     struct ResultPackage
     {
-        std::int64_t split_idx;
-        std::int64_t batch_idx;
+        std::size_t split_idx;
+        std::size_t batch_idx;
         std::string data;
         std::string label_data;
 
         // Calculate size of data in the structure
         std::uint64_t size() const
         {
-            return sizeof(std::int64_t) * 2 + data.length() + label_data.length();
+            return sizeof(std::size_t) * 2 + data.length() + label_data.length();
         }
     }; // struct ResultPackage
 } // namespace apsi
