+++ common/native/apsi/util/utils.h
@@ -52,7 +52,7 @@ namespace apsi
         Compute secure Sender bin size
         */
         std::uint64_t compute_sender_bin_size(
-            std::uint32_t log_table_size, std::uint64_t sender_set_size, std::uint32_t hash_func_count, std::uint32_t binning_sec_level, std::uint32_t split_count);
+            std::uint32_t log_table_size, std::size_t sender_set_size, std::size_t hash_func_count, std::uint32_t binning_sec_level, std::size_t split_count);
 
         extern Stopwatch sender_stop_watch, recv_stop_watch;
     } // namespace util
